dcu: Fix build warnings when CONFIG_FSL_DCU_FB defined
Issue: ALB-2747

+++ drivers/video/fsl_dcu_fb.c
@@ -218,7 +218,8 @@ static int layer_ctrldesc_init(int index, u32 pixel_format)
 		    DCU_CTRLDESCLN_2_POSY(0) |
 		    DCU_CTRLDESCLN_2_POSX(0));
 
-	dcu_write32(&regs->ctrldescl[index][2], (unsigned int)info.screen_base);
+	dcu_write32(&regs->ctrldescl[index][2],
+		    (unsigned int)(uintptr_t)info.screen_base);
 
 	switch (pixel_format) {
 	case 16:
